Fix auto scrolling problem
Add a ref-specific event listener to detect user scrolling in the chatPage div. This is used to cancel the auto scrolling in case the user scrolls. Also added a throttle for the onUserScroll to prevent too many actions in a short period.

Additionally, removed libraries/code related to image uploading. #18